exercise 4.21

+++ controllers/blogs.js
@@ -5,6 +5,7 @@ const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog
+
         .find({}).populate('user', { username: 1, name: 1 })
 
     response.json(blogs)
@@ -22,10 +23,6 @@ blogsRouter.get('/:id', async (request, response) => {
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    console.log(request.token)
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token missing or invalid' })
-    }
     const user = await User.findById(decodedToken.id)
 
     const blog = new Blog({
@@ -43,8 +40,17 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if (blog.user.toString() === user._id.toString()){
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(404).end()
+    }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
